Split git tests into files
